Fix assertion in chooseKRandomIndicesFrom
When verifying that all chosen indices are distinct, they are placed
into a set and the size of the set tested against the size of the
vector, chosenIndices.

However, the type of this set is set<int>, whereas the type of
chosenIndices is vector<int64_t>.

Change the size of this set to set<int64_t> to match chosenIndices.

This fixes a bug where the assertion was tripped erroneously, presumably
because two of the values in chosenIndices, *when cast from int64_t to
int*, had the same value.

+++ my-own-problems/shared/random-utilities.h
@@ -219,7 +219,7 @@ std::vector<int64_t> chooseKRandomIndicesFrom(int numToChoose, int64_t numToChoo
 
     std::sort(chosenIndices.begin(), chosenIndices.end());
     assert(static_cast<int>(chosenIndices.size()) == numToChoose);
-    assert(static_cast<int>(std::set<int>(chosenIndices.begin(), chosenIndices.end()).size()) == numToChoose);
+    assert(static_cast<int>(std::set<int64_t>(chosenIndices.begin(), chosenIndices.end()).size()) == numToChoose);
 
     return chosenIndices;
 }
